chore: migrated magnet tables into a standalone magnet feature folder for seperation of concerns
