+++ lib/database/database.dart
@@ -1,8 +1,9 @@
 import 'package:academia/features/agenda/data/models/agenda_event.dart';
-import 'package:academia/features/auth/data/models/magnet_credentials.dart';
 import 'package:academia/features/auth/data/models/token.dart';
 import 'package:academia/core/data/json_converter.dart';
 import 'package:academia/features/institution/data/models/institution.dart';
+import 'package:academia/features/magnet/data/magnet_credentials.dart';
+import 'package:academia/features/magnet/data/magnet_student_profile.dart';
 import 'package:academia/features/profile/data/models/user_profile.dart';
 import 'package:academia/features/todos/data/models/todo.dart';
 import 'package:academia/features/sherehe/data/data.dart';
