+++ lib/features/magnet/data/magnet_credentials.dart
